WIP Refactor Tasks
